replaced depreciated KVC methods takeValue:forKey:, takeStoredValue:forKey:,  storedValueForKey, handleQueryWithUnboundKey:, handleTakeValue:forUnboundKey:, unableToSetNilForKey:, valuesForKeys: with setValue:forKey:, setValue:forKey:, valueForKey:, valueForUndefinedKey:, setValue:forUndefinedKey:, setNilValueForKey, dictionaryWithValuesForKeys: respectively.  There may be issues with replacing storedValue methods.  No simple equivalent for those. Used NSInteger, NSUInteger where required.

+++ EOAccess/EODebug.m
@@ -50,7 +50,7 @@ BOOL	EORegistrationDebugEnabled = NO;
 
 static void setBoolean(BOOL *boolean, NSString *key,NSDictionary *environment, NSArray *arguments, NSUserDefaults *defaults)
 {
-	int				index = [arguments indexOfObject:EOFormat(@"-%@", key)];
+	NSInteger			index = [arguments indexOfObject:EOFormat(@"-%@", key)];
 	NSString		*value = nil;
    
 	if (index != NSNotFound)
